refactoring features for subscription class
