+++ src/Controller/Gateways/AbstractGateway.php
@@ -100,7 +100,7 @@ abstract class AbstractGateway extends WC_Payment_Gateway
         $this->postFormatter = $postFormatter ?? new PostFormatter;
         $this->model = $gateway ?? new Gateway;
         $this->checkout = $checkout ?? new Checkout;
-        $this->subscription = new Subscription;
+        
         $this->wooOrderRepository = $wooOrderRepository ?? new WooOrderRepository;
         $this->template = $template ?? new Template;
         $this->id = 'woo-pagarme-payments-' . $this->method;
@@ -121,11 +121,7 @@ abstract class AbstractGateway extends WC_Payment_Gateway
         add_action('woocommerce_receipt_' . $this->id, array($this, 'receipt_page'));
         add_action('woocommerce_thankyou_' . $this->id, [$this, 'thank_you_page']);
         add_action('admin_enqueue_scripts', array($this, 'payments_scripts'));
-
-        //Subscriptions
-        if ($this->hasSubscriptionSupport() && $this->subscription->hasSubscriptionPlugin()) {
-            $this->addSupportToSubscription();
-        }
+        $this->subscription = new Subscription($this);
     }
 
     /**
@@ -136,36 +132,6 @@ abstract class AbstractGateway extends WC_Payment_Gateway
         return false;
     }
 
-    /**
-     * @return void
-     */
-    private function addSupportToSubscription(): void
-    {
-        $this->supports = array(
-            'products',
-            'subscriptions',
-            'subscription_cancellation',
-            'subscription_suspension',
-            'subscription_reactivation',
-            'subscription_amount_changes',
-            'subscription_date_changes',
-            'subscription_payment_method_change',
-            'subscription_payment_method_change_customer',
-            'subscription_payment_method_change_admin',
-            'multiple_subscriptions',
-        );
-        add_action(
-            'woocommerce_scheduled_subscription_payment_'.$this->id,
-            [$this, 'schedule_subscription_payment'],
-            10,
-            2
-        );
-    }
-
-    public function schedule_subscription_payment($amountToCharge, $order)
-    {
-        $this->subscription->process($order);
-    }
 
     public function payments_scripts()
     {
